requirements.txt
Django==4.1.3
Pillow==9.5.0
djangorestframework==3.11.1
drf-yasg==1.20.0
django-cors-headers==3.5.0